sw_redlinehide: add an indirect iteration mode to SwIterator
There are lots of existing loops over SwTextFrames of SwTextNodes,
and these need to keep finding the frames if there's a
WriterMultiListener in between.

Add a template parameter to SwIterator that can be used to make the
iteration transparent, i.e. return the client of the
WriterMultiListener instead of the WriterMultiListener itself.

Change-Id: I8b67a027b392b08408f5fbf26be22884384aa6cf